Extract helpers from test for readability.
